refactor(ui): add totp-first fork-ui module paths
