+++ packages/fork-ui/src/index.ts
@@ -8,17 +8,17 @@ export function wrapRoutes<T>(routes: T): T {
 
 export { LoginApp } from "./login"
 export { BootstrapSignupApp } from "./bootstrap-signup"
-export { TotpApp, TwoFactorApp } from "./two-factor"
-export { TotpSetupApp, TwoFactorSetupApp } from "./two-factor-setup"
+export { TotpApp, TwoFactorApp } from "./totp"
+export { TotpSetupApp, TwoFactorSetupApp } from "./totp-setup"
 export { PasskeySetupApp } from "./passkey-setup"
-export { ManageTotpDialog, ManageTwoFactorDialog } from "./manage-2fa-dialog"
+export { ManageTotpDialog, ManageTwoFactorDialog } from "./manage-totp-dialog"
 export { PasskeyManagerDialog } from "./passkey-manager-dialog"
 export { SessionIndicator } from "./session-indicator"
 export { SessionExpiredOverlay } from "./session-expired-overlay"
 export { SettingsAuthenticationSection } from "./settings-authentication-section"
 export { SettingsAuthSessionTab } from "./settings-auth-session-tab"
 export { SettingsAuthPasskeysTab } from "./settings-auth-passkeys-tab"
-export { SettingsAuthTotpTab, SettingsAuthTwoFactorTab } from "./settings-auth-twofactor-tab"
+export { SettingsAuthTotpTab, SettingsAuthTwoFactorTab } from "./settings-auth-totp-tab"
 export { SettingsAuthFooterLogout } from "./settings-auth-footer-logout"
 export { SettingsAuthProvider, useSettingsAuth } from "./settings-auth-state"
 export { SecurityBadge } from "./security-badge"
