+++ packages/fork-ui/src/manage-totp-dialog.tsx
@@ -1,0 +1,1 @@
+export { ManageTotpDialog, ManageTwoFactorDialog } from "./manage-2fa-dialog"
