+++ packages/fork-ui/src/manage-totp-panel.tsx
@@ -1,0 +1,1 @@
+export { ManageTotpPanel, ManageTwoFactorPanel } from "./manage-2fa-panel"
