+++ packages/fork-ui/src/settings-auth-totp-tab.tsx
@@ -1,0 +1,1 @@
+export { SettingsAuthTotpTab, SettingsAuthTwoFactorTab } from "./settings-auth-twofactor-tab"
