+++ packages/fork-ui/src/totp-setup-flow.tsx
@@ -1,0 +1,6 @@
+export {
+  TotpSetupFlow,
+  TwoFactorSetupFlow,
+  type TotpSetupBootstrap,
+  type TwoFactorSetupBootstrap,
+} from "./two-factor-setup-flow"
