+++ packages/fork-ui/src/totp-setup.tsx
@@ -1,0 +1,1 @@
+export { TotpSetupApp, TwoFactorSetupApp } from "./two-factor-setup"
