+++ packages/fork-ui/src/totp.tsx
@@ -1,0 +1,1 @@
+export { TotpApp, TwoFactorApp } from "./two-factor"
