+++ packages/fork-ui/src/two-factor-setup.tsx
@@ -1,4 +1,4 @@
-import { TotpSetupFlow, type TotpSetupBootstrap } from "./two-factor-setup-flow"
+import { TotpSetupFlow, type TotpSetupBootstrap } from "./totp-setup-flow"
 
 declare global {
   interface Window {
